Add CollectionsOverview component and move CollectionPreview renders inside it.
